requirement.txt
streamlit==1.30.0
tabula-py==2.9.0
pandas==2.0.3
